backend/requirements.txt
django==5.1.5
markupsafe==3.0.2
pyyaml==6.0.2
sqlalchemy==2.0.37
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.1
black==25.1.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cohere==5.13.11
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
docopt==0.6.2
et-xmlfile==1.1.0
fastapi==0.115.6
fastavro==1.10.0
filelock==3.17.0
flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-text-splitters==0.3.5
langsmith==0.3.4
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pgvector==0.1.6
pillow==11.1.0
pip==25.0
pipreqs==0.4.13
platformdirs==4.3.6
propcache==0.2.1
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.1
pypdf==5.2.0
pypdf2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.66.5
transformers==4.48.2
types-requests==2.32.0.20241016
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
werkzeug==3.1.3
yarg==0.1.10
yarl==1.18.3
zstandard==0.23.0
autocommand==2.2.2
backports.tarfile==1.2.0
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
tomli==2.0.1
typeguard==4.3.0
wheel==0.43.0
zipp==3.19.2